fix: enable fractional crypto + add notional support for dollar-based orders (#247)
Two related Alpaca capabilities the platform was advertising but not
actually wired up.

## 1. Fractional crypto silently dropped (the real bug)

`feedback_kah_crypto_whole_units.md` flagged that `POST /trades` with
`contracts: 0.001` returned 202 but no trade row appeared. The
mechanism: `TradeRecord` had `field :contracts, :integer`, Postgres
truncated 0.001 → 0 on insert, `validate_number(:contracts,
greater_than: 0)` rejected the changeset, the worker logged an error
that nobody saw, the API had already returned 202.

Fix:
- Migration alters `trade_records.contracts` from `:integer` to
  `:numeric` (widening, no rewrite, existing rows preserved).
- Schema field becomes `:decimal`. All consumer call sites already
  wrap in `Decimal.new(...)` which accepts both integer and Decimal,
  so no consumer changes needed (Trading FIFO PnL, vault ABI,
  collective intelligence — all keep working).
- API serializer coerces the now-Decimal `contracts` back to a JSON
  number (integer when whole, float when fractional) so the agent
  prompt + MCP `get_trades` consumers don't see a sudden string.

## 2. `notional` parameter (USD-based fractional / dollar orders)

Per Alpaca docs, `/v2/orders` accepts either `qty` OR `notional`
(mutually exclusive). Notional unlocks "spend $250 on AAPL" without
the agent computing share counts at the current price.

Wiring:
- `trades_controller` accepts a top-level `notional` field, validates
  exactly one of `contracts`/`notional` is present, threads `notional`
  into the worker's args.
- `TradeExecutionWorker` stores agent-supplied notional verbatim, and
  estimates `contracts` from `notional / fill_price` so the schema's
  required field is populated. The settlement worker rewrites the
  estimate with the actual broker qty when the fill comes back.
- `AlpacaClient.order_body/4` sends `"notional"` and omits `"qty"`
  when notional is set. Options always fall through to qty (Alpaca
  rejects notional on options).
- Agent prompt updated with notional rules and a note that crypto
  is fractionable.

## Tests

- 4 new TradeRecord changeset tests — fractional, integer, zero,
  negative — all pass.
- 4 new AlpacaClient order_body tests — notional sent, qty omitted
  when notional present, options ignore notional, crypto fractional
  qty passes through.

Verified the existing `prompts_have_single_quotes?` invariant still
holds — the prompt edit uses "Do not" instead of "Don't" so the
`codex '<prompt>'` shell wrap stays intact.

+++ lib/kite_agent_hub/trading/trade_record.ex
@@ -16,7 +16,12 @@ defmodule KiteAgentHub.Trading.TradeRecord do
     field :market, :string
     field :side, :string
     field :action, :string
-    field :contracts, :integer
+    # `:decimal` (was `:integer` until 2026-05-01) so the schema can
+    # represent fractional contract sizes for crypto + future fractional
+    # equity orders. The previous integer field silently truncated
+    # 0.001 BTC to 0, which then failed validate_number(greater_than: 0)
+    # and dropped the trade with no row inserted.
+    field :contracts, :decimal
     field :fill_price, :decimal
     field :notional_usd, :decimal
     field :status, :string, default: "open"
